requirements.txt
numpy==1.22.4
llvmlite
setuptools==59.5.0
numba
tensorboardX
easydict
pyyaml
scikit-image
tqdm
open3d
wandb==0.13.1
matplotlib 
pandas 
plyfile 
pyaml 
pyopengl 
PyQt5 
pyqtgraph 
scipy 
scikit-learn 
PyMieScatt 
pyquaternion
spconv-cu111==2.1.21
#SharedArray==3.1.0